ARC-1997: add explanation to the class why we do this; moved to separate file
